container-lsnuyl4-main/src/requirements.txt
Flask
SQLAlchemy
gunicorn
psutil
PyMySQL
coloredlogs
python-dotenv
ollama
requests
